All SPs are broken because I pass args now.
Not sure if I am even keeping this branch, since it might
be better to treat this as a complete design and to start again
more carefully on monday. I would prefer to write a new venture
to prototype this in an easier language, with less to break and
maintain.