Contorting around Pythonic incompetence.
Turns out the multiprocessing module uses pickle rather than fork to
send jobs to the workers, so all the arguments to Pool.map (including
the function !?) need to be picklable.

Turns out dynamically-created named tuples are not picklable, so can't
be sent over the wire.

Turns out regular dicts are picklable, but can't be used as keys to
other dicts, because Guido doesn't trust me not to mutate them.

Turns out there is no standard frozen dict class that would act like a
dict except for being immutable and therefore hashable.

Also turns out that the order of the keys matters.

So we use ordered dicts first, which are picklable; then convert them
into named tuples for use as hash keys.