Added support for (sloppy) slice to lite
Note: I refer to this as 'sloppy' slice because it does
not mix over the rest of the entropy in the scaffold,
and simply estimates the density based on a single sample
of said entropy. It is not trivial to implement the correct
version in lite because we currently do not support setting
the random seed. It still behaves correctly on all simple
scaffolds with one principal node and no other entropy.