Set up for automated testing.
Man, trying to make third-party libraries in MIT Scheme sucks!  I need
to wrap test-manager in a special environment so that the client's
redefinition of apply doesn't interfere with it.

Then again, if I put the (let ((client-environment ...)) ...) block
into the load file of the test manager package itself, then I could
hook a documentation generator tool to the export list (but macros
don't carry their environments of definition?)