First draft of new church prime parser.
Not hooked up to anything yet, but the intent is that this will be a
drop-in replacement, including the broken source position records.

Split into an NFA lexer, written with plex, and a LALR parser,
generated by lemonade.  The unambiguous context-free grammar is in
grammar.y and is language-independent -- you can just read it, with
no in-line Python or C code for semantic actions, to see the language
grammar.

I tried to express the same lexical syntax and grammar that the old
parser implements, but the old parser is so tangled a mess of pyparsing
spaghetti that it's hard to say for sure.

I'm including the lemonade-generated grammar.py until I find time to
hook it into the build system.  Until then, if you make changes to
grammar.y, please make sure to regenerate grammar.py with lemonade and
include that in the same commit.