Fix Puma crash when returning a foreign procedure to Python
The issue arises when the aux is null (as auxes of builtin procedures
are) but is assumed to be an instance of ForeignLiteSPAux (which is
created for returned SPs that actually have auxes). Make the
ForeignLiteSP check for null SPAux before translating to Python, and
on the Python side, ask the SP to make a real stack dict when actually
extracting the value.