Environments: don't remove symbol definitions from outer environments
When we bind a symbol in the "global" environment, we actually walk up a tree
of environments to find where it's already bound, and delete it from there.
This doesn't allow for variable shadowing, so instead, remove a binding only
when we redefine a symbol from within exactly the same env frame where it's
already bound.

Testing Done: test/regressions/test_variable_shadowing.py
Bug Number: 167

Signed-off-by: Eli Sennesh <eligottlieb@gmail.com>