Test consistency of the suff stat poisson assessment.
In the case of the Poisson distribution, it is not possible to report
the probability of a given data sequence from its sufficient
statistics, so make_suff_stat_poisson reports the probability of the
statistic (rather than the sequence that gave rise to it).  Therefore,
the raw assessment of make_suff_stat_poisson will differ from the
assessment of the native version; however, their ratios (differences
in log domain) will remain the same.