Fix Gibbs wrong stationary distribution in the presence of brush.
This fixes issue #207 by conforming to Neal's Algorithm 8, which is
justified as an auxiliary variable Gibbs step. The main difference is
that if there are random choices y downstream of the variables being
enumerated, z, we retain the current values of y rather than
resampling new ones for the particle corresponding to the current z.
(The intuition is that randomly sampled values of y will tend to have
lower likelihood than the existing values, so the existing values may
otherwise be discarded more often than they should.)

A possible consequence is that the egibbs operator will consider fewer
different values of the downstream choices than before, but those
downstream choices can be subject to further inference anyway.