What a crock infer (endloop) turned out to be!
The problem was the SIVM's continuous inference pauser: when CI is
paused, it can't be turned off, because it will be resumed.  When
executing a bare ripl.stop_continuous_inference(), this is worked
around by not pausing the continuous inference.  But that doesn't work
for infer (endloop), because the pause is done before starting to
execute the 'infer' instruction, and by the time the 'endloop'
inference SP reentrantly executes stop_continuous_inference, the
resumer is already on the stack.

Hacked around it by maintaining an explicit stack of pending CI
resumers and disabling the top one.