Flush the profiler_configure instruction.
Replace with direct method calls to turn profiling on and off, since
that's the only thing it used to do.

Functional change: enabling, disabling, and querying the profiler's
run state will not interrupt continuous inference.  There is a chance
that this will be problematic "because threading", but I think it
sufficiently unlikely that keeping the current structure is not a good
response to that potential problem.