Rework multivariate normal logpdf code.  Use it in gp.
Compute the Cholesky decomposition of the covariance matrix instead
of computing matrix inverses, if we can; otherwise fall back to
computing matrix inverses.

XXX It would be better to compute the LDL decomposition instead, but
neither numpy nor scipy has code to compute it or to solve systems
with it.