Add part of a two-sample low-covariance test from matrixGP.
We assert that from 200 samples at each of two input points with low
covariance, a GP should pass Pearson's r^2 test of independence,
i.e. fail to reject the null hypothesis that the sample sets are
independent.

This does not add the corresponding high-covariance test, which would
reject the null hypothesis that the sample sets are independent at
each of two input points with high covariance.  Currently the Venture
statistical tests are not able to express this sense of the test.