Tweak FixedRandomness in the new deterministic world order.
FixedRandomness -- which should really be called RandomnessFixer --
is a device for saving and restoring the value of a particular PRNG
state.  It doesn't make sense to default to a newly constructed PRNG
state -- which would be constructed once per process and used only
for the arguments of FixedRandomness, because that's how Python
optional argument defaults work.  Rather, every caller should have a
particular PRNG state in mind, so make those arguments mandatory.

Propagate the change to the tests and to BogusArgs.  For now, we let
BogusArgs draw from system entropy, until we push that into the test
harness itself.