Oops.  Tag as statistical the actually random portion of the test body.
Before this, testForEachParticleIsIndependent would draw a
distribution that was supposed to be Gaussian, and then do a
'statistical test' that would repeatedly deterministically check that
sample for Gaussianness, and complain if the aggregate p-value was too
low.  This caused an excess of chance failures -- the effect of the
Fisher Method in this circumstance is to substantially lower the
actual statistical significance of the reported result.

The fix should restore the test to the expected significance level,
and thereby cause it to spuriously fail less often.