Narrow the scope of Lite's bizarre "normalization" of block and scope IDs.
The code's type discipline simplifies.  The external change is that
various methods no longer accept raw Python objects as scope or block
IDs, demanding VentureValue objects instead.  The extant affected
clients are a couple tests that used the numBlocksInScope and
numNodesInBlock functions.

The narrowing also has the effect of distinguishing ordered range
block specifiers from block ids that are lists, permitting Lite to use
the latter as ids as well as Puma does.