Basic procedure applications working
Import all the lite SPs and wrap the "simple" ones in the new
interface (what little of it exists for now). Also add an SP registry
like Lite's, since we'll want it eventually.

Now we're passing test.core.test_basics.testTriviality{1,2},
testFlip{1,2}, and several of the modules in test.conformance.sps.