Working observe
Constrain is an SP method now, so procedures can override it to do some
custom processing. By default, fall back to absorbing via logDensity.

When constraining, check whether the node has any children other than
the source of the constraint; bail if so. This is stricter than
Lite (which allows non-random non-requesting children), but I don't want
to have to think about it yet.