Add a mem re-entrancy guard.
This guards against programs like

  assume f = mem(proc (x) {
    if (flip()) { 1 }
    else { 1 + f(x) }
  }

where a memoized procedure calls itself with the same arguments,
producing inconsistent results that overwrite each other in the mem table.

The existing Lite and Puma backends guard against this by checking if a
mem table entry appeared during evaluation of the procedure call. This
detects the consistency bug, but only after the recursion starts
unwinding.

Here, we instead populate the mem table entry with a sentinel value, and
check for the sentinel on entry to the procedure (rather than checking
on exit). This leads to similar results, but fails slightly faster
because the recursion doesn't happen in the first place.