Add unconstrain.
All of test.stack.functionality.ripl_test passes now except for
test_force.

Force is broken because the new specification of unconstrain is that it
resimulates the no-longer-constrained value from the prior, so that
sequences of constrain and unconstrain don't mess up the stationary
distribution. Unfortunately, force was relying on the behavior that
unconstrain would leave the previously-constrained value there. I'm not
sure what to do about this.