Add restore and getState/setState methods
to the interface, to facilitate undo/restore functionality.

The unapply method is now expected to use args.setState to store any
information needed by restore, which can be retrieved with
args.getState. In the case of SimulationSP, the state is just the
simulated value.

Also rewrite SimulationSP.constrain, which used to be defined in terms
of unapply; now that unapply calls setState, this definition doesn't
work anymore (or else constrain calls setState, which doesn't make much
sense and crashes in the OmegaDB). Instead, inline the old definition of
unapply into constrain. For symmetry, inline apply into unconstrain too.

As of this commit, these methods aren't yet called, and
getState/setState don't have implementations.