Make requests work in the new environment
Add a TraceHandle class that provides the requesting interface. For now
this is conflated with the old Args object but they can be separated
later.

Also, for now, the untraced interpreter will store the results of every
request, even non-repeatable ones (like ordinary compound procedure body
evaluations), because of the way that request results are retrieved with
requestedValue. Later it should be possible to change it so that the
requesting interface returns the result value itself, and it is retained
only for requests with repeatable request_ids, like those of mem.