Introduce ~, <~, and ~> as aliases for =, <-, ->, respectively.
Why?  Because ~ semi-commonly means "sample from" in statistics
papers.  As an alias this allows us to write code that will be more
familiar to that audience.  Perhaps we will eventually add something
like a type system that will actually enforce an interesting semantic
distinction between the squiggly and hard binders.