Teach the pylint config to be sensitive to the installed pylint version.
Why?  Because pylint configuration files, at least in blacklist style,
are not compatible across versions.  To wit, if pylint introduces a
warning in a new version, it will be on by default.  If you then put
it in the configuration file to suppress it, the old version of pylint
will fail to parse the file.