Separate the list of inputs from the no-longer-Args.
The trace handle no longer attempts to satisfy Lite's IArgs
interface. In fact, it's now only responsible for general interfacing
between an SP and its surrounding trace; all the information specific to
the calling context (application id, list of input arguments) is passed
separately and explicitly.

For now, the inputs are provided as a list of nodes, whose values can be
retrieved by node.value. A TODO item is to abstract this and require the
trace handle to get the value out of the input node. (The input nodes
are necessary, not just the values, because SPs are allowed to put them
in environments for requests, and the nodes carry the information
required to trace the dependencies.)

Another TODO item is to add an initialize method, and have the trace
handle be provided to the SP there, so that it doesn't need to be passed
to every single method. Currently this is blocked by the fact that the
trace handle is context-dependent, which in turn is because
trace_handle.new_request actually means "restore the request" when we're
in a restoring context. This can be fixed by adding a separate
trace_handle.restore_request method and making SPs responsible for
calling that appropriately.