Stop deep-copying the trace on every operation.
Originally I did this because I was afraid of what would happen if I
let trace actions mutate their input trace when the trace action was
itself evaluated in a trace (and therefore could be unevaluated and
replayed in a different order).

Now, I think it is safe as long as:

1. Mutable trace variables are "single-use": if a trace is passed
directly to an inference action's action_func, it must not have any
other direct children in the dependency graph. This is ensured by
`run_in` making a copy of the input trace (so that any mutations aren't
visible to other things that read the same trace variable) and `bind`
following this discipline internally (where the second action is invoked
on the trace returned by the first action, rather than on the input
trace directly).

2. Every primitive trace action defines a method for undoing itself, so
that `run_in` can be unapplied. (Adding these for the existing trace
actions is a TODO.)