Add an experimental make_exchangeable_sp example.
This is a concept for an alternative method for defining SPs, where an
SP's apply method is composed from some primitive invertible trace
actions, and its unapply method is automatically given by inverting the
component primitives. This is appealing both interface-wise, because it
reduces the amount of redundant code needed to define an SP, and
implementation-wise, because it allows make_exchangeable_sp and
procedures defined using it to be evaluated directly in the same ambient
trace where they are used, which simplifies issues of scoping and
marshalling across traces.

This example is not fully implemented because the primitives don't
actually have inverses defined in the code, but hopefully it is enough
to illustrate the idea.