Work around an exchangeable coupling bug in lightweight regen.
As written, x3 is proposed conditioned on x1 and x2, but not on x4 and
x5; and the likelihood weight for x4 and x5 given x3 is not accounted
for automatically (because the args are the same, so lightweight regen
assumes the weights are the same, but actually the aux is different so
the weights are different).

Work around this by putting constrained kernels on x4 and x5 manually,
which forces their weight to be included.