Add a fourth CRP inference program example.
Interestingly, this should be equivalent to example2c: instead of
keeping x4 and x5 constrained and accumulating the weight (which is used
to accept/reject), it just resimulates x4 and x5 (and tests them against
the observed values to accept/reject). Assuming the simulator and
assessor agree, the rejection rates and result distributions should be the same in both cases.