Added the trace (linear algebra) for a matrix.
Added the trace of a matrix as a simple SP. Trace is defined in the linear
algebra sense, i.e. the sum of the elements of the main diagonal of a square
matrix.

Caveat: the input type for this is SP is matrix. I understand that there is no
type that enforces a matrix that is square.

Purpose of this SP: help to implement GPs natively in mite.