Consistent interpretation of a subproblem.
The subproblem should include all nodes where kernels are being
applied. Unlike before, all nodes not mentioned are assumed to be
completely uninvolved, i.e. their values and all their parents' values
are unchanged. These nodes are neither extracted nor restored.

This is a step toward toward subproblems and regen having the same
probabilistic meaning, independent of the structure of the trace.

For now, the mem example is broken. I plan to make several more changes
that are likely to touch it, so I will fix it later.