Stub regen for the dependency graph.
Currently this doesn't do its own traversal, and relies on the fact that
the scaffold constructed by single_site_subproblem is actually in
forward evaluation order. However, the processing for each node should
be basically right.

Note: piggybacking off single_site_subproblem is not a long term
solution because single_site_subproblem traverses the whole program,
which defeats the asymptotic performance gains from using the dependency
graph. But on a simple example, it seems the subproblem traversal is
much cheaper than the regen traversal for some reason.