Tweak the requesting interface to return Node(addr, value)s.
The intent of this change is that simple requesting SPs can get the
result value directly from what is returned by new_request, rather than
it being recorded in the trace and then having to be immediately
retrieved. This way, the untraced trace doesn't have to record results
every compound procedure (which is a major memory leak).

Summary of the change:

- new_request and restore_request now return Node(addr, value).

- free_request now returns nothing, because the return value
  wasn't being used anyway.

- value_at now takes a Node(addr, value) and returns the value. (This is
  preferred over accessing .value directly, because later the trace
  might override value_at(node) -> value.

- get_request takes the request_id of an existing request and returns
  the Node(addr, value) corresponding to that request. This is used by
  procedures like mem that have persistent/shared requests. Hence,
  value_at(get_request(...)) is equivalent to the old value_at(...).