Add some low-level trace programming operations.
This makes it possible to write low-level inference programs that do not
rely on Venture's utilities for working with inference
subproblems. While most users would not want to use these directly, they
are useful for representing what extract/regen on a subproblem does
behind the scenes, and potentially useful as a representation for a
future "subproblem compiler". Also, with a few additional operations, it
should become possible to write eval and regen themselves as inference
programs (like the metacircular evaluator).