Prototype a constructor for tail-assessable SPs.
This is called proc_, because proc is currently a keyword; it wants to
be called proc, because vkm has the idea that tail assessability is what
distinguishes procs from lambdas. (A lambda is just a bare procedural
abstraction, with no metaprograms of its own, whereas a proc can
implement constrained kernel by delegating to the constrained kernel of
the SP in its tail position.)

The goal is that there can be many of these, which implement various
analyses on the source of the procedure body.