Rename MinimalScaffold -> Scaffold
There's nothing particularly minimal about it*, and it's confusible with
the notion of "minimal subproblem" (i.e. an operator that finds the
subproblem that resimulates the least while being able to evaluate a
given proposal).

* anymore; the original naming was because it was "minimal" in the sense
that it just stores an ordered map of kernels rather than a bunch of
distinct node sets, but that doesn't seem important anymore.