Rename VentureTangentFunction to VenturePartialDiffableFunction.
Elaborate to compute partial derivatives with respect to the
fixed/closed-over parameters or one input parameter, and to compute
the pushforward -- i.e., to compute the value at each input and the
derivative that input simultaneously.

WARNING: This breaks GPs, which need to be updated for this API.
WARNING: Not yet tested!