Update GPs to support gradientOfLogDensity.
Start reworking covariance kernels to support partial derivatives.

For now, for gp([x_0, x_1, x_2]), this will compute the partial with
respect to x_0 first, and then the partial with respect to x_1 next,
and so on, since vectorizing it is too miserable to contemplate at the
moment.

Also use uppercase letters for arrays of inputs and lowercase letters
for individual inputs in covariance.py, since for now we will be
computing partial derivatives at {x_0} x {y_0, y_1, y_2, ..., y_{k-1}}
in order to avoid having to stray into tensor territory.

WARNING: This is missing a little bit of testing, and by a little bit
I mean the concept works on a whiteboard.