Treat d/d(r^2) bump(r^2) at 0 as zero.
The bump kernel is not actually differentiable at zero, i.e. on the
line x = y, but the limit as the distance goes to zero of its
derivative is zero, and the restriction of the map (x, y) |--->
e^{-1/|x - y|} to the line x = y is differentiable, and we are really
only interested in covariances of points outside the line x = y, or
in self-covariances/variances of points on the line x = y.  So while
we haven't formalized distinguishing covariances from
self-covariances/variances, this will do for now.