Rename GP covariance kernel bump ---> deltoid.
This is not actually a bump function.  It is, however, supposed to be
a continuous approximation to a delta kernel, and it resembles (but
is not exactly) a sigmoid curve -- hence `deltoid'.

Note that the tolerance parameter is hokey; what is currently called
t is actually tau^1/s, where tau is a sensible tolerance parameter
scaling the unit where the middle of the drop-off is, and s is the
steepness.  But changing this requires recomputing the derivatives,
and this kernel is not a priority right now.