This is what I get for trying to write Haskell in Python.
I thought I set up an invariant in which I had an algebraic type in
this code that was either a node, or an OrderedFrozenSet of nodes, or
a SamplableMap of things.  But instead of tagging that algebraic type
myself, I relied on Python's tags.  And lo and behold, when this code
is applied in the context of enumerative Gibbs, my node sets are
suddenly not all OrderedFrozenSets, but some of them are PSets
instead.

XXX Hack: Generalize/canonicalize on the paths that are actually
exercised by the existing test case.

Fixes #634, but not unreasonable to reopen if I didn't catch all the
mistakes of that type.