Replace bogus_mapv by parallel_mapv.
New implementation of parallel_map creates as many workers as there
are CPUs, sharing master's memory with fork to obviate the need for
pickle.  (Otherwise we could just use multiprocessing.Pool().map
instead of parallel_map.)

This is not a long-term solution!  It is a stop-gap for a student
emergency.  In particular, we should not encourage indiscriminately
using as many workers as there are CPU cores -- e.g., consider doing
it with resample_multiprocess, so that if you have 64 cores you'll
end up with 4096 workers in total, and waste a lot of time in
scheduling.