Fix Ulli's mystery hmc type error.
Grump.  The problem was that hmc(subproblem, epsilon, ...) would fail
to unwrap the epsilon argument, thinking it might be the block part of
a scope block pair.  There is actually no way to deal with this,
except to suppress all unwrapping of arguments for all operator SPs.

Puma is not affected because it already unwraps all incoming Venture
Values through their stack dicts.