Commented out some of the sanity asserts in mvnormal.
Structure learning happens to sometimes propose structures with lots
of multiplications that will inevitably result in NAN values when
logpdf is computed. I have therefore decided to remove assert statements
for such cases - instead return -inf for logpdf - hoping that such proposals
will simply be rejected. This is very hacky at the moment and should NOT be merged
into Master.