Enumerate values after detach.
This gives the SP a chance to correctly identify what it might return
once some applications have been unincorporated.  If unincorporating
a CRP application emptied a table, then we would like to return that
table as a possibility, but not another empty table.  However, if we
enumerate values before unincorporating the applications, the CRP
will enumerate both tables.

Enable an automatic test for this.

Second part of fix for Github issue #470.