Abstract the interesting content of the random_site definition into the prelude.
Sadly, one piece of boilerplate remains, because it is not possible to
insert a make_elementary_sp definition into every trace as a prelude
item.

Why not?  Because make_elementary_sp makes a trace internally, to
serve as the trace in which the methods of the SP are evaluated.  It
is therefore necessary to delay the evaluation of the
make_elementary_sp expression.

In principle it might be possible to eliminate the boilerplate by
introducing something like lazy symbol bindings which get evaluated
the first time they are looked up.  It might also be possible to
eliminate this particular boilerplate with a module system, where
random_site needed to be imported (and thus would not be constructed
in every private trace of every make_elementary_sp call).