Correctly handle *non*-shadowed, non-familied makerNodes as well.
Introduce hasMadeSPRecordAt in order to avoid violating an
abstraction boundary.  For now, nobody should be asking this of
Particles -- only Particles should be asking it of Traces.  If this
changes, you will be loudly notified and you can change the
assertion.

Use hasMadeSPRecordAt judiciously to decide whether we can ask the
base trace about the made SP families at a node, if the derived
particle has them but we're not yet sure whether it was shadowed /
inherited or not.