Introduce the concept request_constraint_kernel; adjust scaffold construction to use them.
Like propagating_kernel is a special kernel that looks like the
resimulation of the OutputPSP, request_constraint_kernel is a special
kernel that looks like absorbing at the RequestPSP (without specifying
what the OutputPSP will do).

This suffices to unbreak the example of changing the argument to a
compound procedure.