WIP updated the description of inference and summary of status (originally written as DM to @fsaad in Slack):
I have pushed a notebook and two plugin-files to Venturecxx branch _20161030-schaechtle-stable-gp-inference_ , please find them in the directory `venturecxx/examples/inverse_compilation`.

I will not push the airline data - please DM me for a version and please note that I use normalized data. I don't want to push it because I don't know my rights wrt to redistribution of this data.

*Important*: I use `resample_multiprocess`, thus, when you initialize the notebook server make sure that this flag is set:
```
OPENBLAS_NUM_THREADS=1
```
If you don't set this flag you will leak multiprocessing - BLAS will call all available cores (i.e. 64) - for each of the 60 chains that run in parallel. That causes the process to deadlock themselves.

I have re-run an experiment over night using the airline data and the notebook that I pushed (using a very similar style to what is presented in the paper - but for inference using `mh(default, one, 5000)`, the extrapolation performance looks pretty good (see final plot) so I assume that this works. The experiment took roughly 3 hours. I have started a similar experiment with gradients -  for some reason, this did not finish yet (after 13 hours). I am worried that I am leaking multiprocessing again because htop's output looks suspicious. The cores are not continually busy but go to zero % every now and then. Also, not all the process are reported for the same amount of time.

Finally, I have taken a look at the posterior on structures and they tend to be more complicated than in earlier experiments (with lot's of multiplications and lot's of constant kernels). One explanation for that is that I changed (relatively late into the submission process) the gradient-safe
uniform prior to the gradient-safe exponential prior. This means that the prior is less likely to produce large hyperparameter values and thus the structure generator is producing some of those larger hyperparamter values by scaling kernels with multiplying with a number of constant kernels. I think we really need a proper simplifier - that is full blown -simplifying any overly complicated structures.

Please let me know if you have any questions or run into any crashes.