WIP initial version of the simplest  simplifier I could think of.
There are three pattern matching rules encoded in this framework, all only
working on products of kernels (for now).I will test this using a slightly
modified grammar, namely one that only spits out sums of products as source
code. This should be equally expressive in terms of structures a grammar can
can produced compared to the grammar I have now but it makes simplification
significantly easier.

For now, the simplification of products is baked into Venture as an SP. I did
this for rapidly creating a prototype.In future I plan to re-implement all
of this code (or better, the code that lasts) in VentureScript.

This is work in progress and not great code at all. A also added a few automated
tests for this to the Venture test suite.

The immediate next step is to code  up the modified version of the grammar (I
think I only need to change a few lines) so that it produces sums of products
exclusively. Then, I will run a notebook to produce some anecdotal evidence
for simplification being effective.