Change ssrmh_step to return the new trace on accept,
so that the current state of the Markov chain is not polluted with
random choices that do not exist in the execution it represents.

This fixes a bug: if such choices are present, one may be chosen as
the site.  That would be poor, because the new trace would not have a
value there at all, which would cause a crash when ssrmh_step would
try to extract that value in order to set the state trace's value in
that location to it.