Migrate change point kernel from gp-inverse-compilation repo to Ventuercxx.
Originally, I planned to have change point kernel as a foreing SP somehow. However,
that caused a problem when I tried to save the entire ripl to disk. Thus, I migrated
the change point kernel operator to Venturecxx.

As of yet, I haven't migrated the tests for this kernel. Mainly because they use pytest
and depend on other code in the gp-inverse-compilation repo.